requirements.txt
notion_client==2.0.0
PyInquirer==1.0.3
PyYAML==6.0
